Each ZHigh op goes to a seperate file
